% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_CT.R
\name{analyze_CT}
\alias{analyze_CT}
\title{Analyze CT values}
\usage{
analyze_CT(
  data,
  col = "CT",
  subset = c("Calibrator", "Reference"),
  group = c("Detector"),
  funs = c("mean", "mean_centered", "mean_substract", "sd", "error"),
  new_col_name = NULL
)
}
\arguments{
\item{data}{A tibble with qPCR data.}

\item{col}{A character string. The column to apply the function to.}

\item{subset}{A character string. Select a subset, either \code{Calibrator} or
\code{Reference}.}

\item{group}{A character vector. The groups to consider when applying the
function on the column.}

\item{funs}{A character string. Select a function to apply to the column,
either \code{mean}, \code{mean_centered}, \code{mean_substract}, \code{sd}, or \code{error}.}

\item{new_col_name}{A character string. Select a name for the new column. If not
specified the new column will be named according to the original column \code{col}
and the function \code{fun}.}
}
\value{
A tibble with a column containing the calculated values.
}
\description{
This function applies a function in a subset of the data (e.g. the references
assays or the calibrator sample) and performs a calculation with this value
on the remaining dataset and stores it as a new column. It is sensitive to
groups. This is useful for many calculations in the analysis of qPCR data.
}
