% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_CT.R
\name{analyze_ddCT}
\alias{analyze_ddCT}
\title{Analyze qPCR data with ddCT method}
\usage{
analyze_ddCT(
  data,
  sample = "Sample",
  detector = "Detector",
  ct = "CT",
  calibrator,
  reference,
  simplify = TRUE
)
}
\arguments{
\item{data}{A tibble with qPCR data.}

\item{sample}{A character string that identifies the sample column.}

\item{detector}{A character string that identifies the assay/detector column.}

\item{ct}{A character string that identifies the column with the CT values.}

\item{calibrator}{A character string with the name of the calibrator sample.}

\item{reference}{A character string with the name of the reference assay.}

\item{simplify}{A boolean value that specifies the returned data tibble.
If \code{FALSE} it will return intermediate dCT values.
If \code{TRUE} it will return only ddCT values.}
}
\value{
A tibble with ddCT values.
}
\description{
This function analyzes a qPCR dataset in \code{data} according to the ddCT method.
}
