% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melting_Analysis.R
\name{assay_positions}
\alias{assay_positions}
\title{Extract the detectors from a csv files with melting data.}
\usage{
assay_positions(file_name)
}
\arguments{
\item{file_name}{The file to load from.}
}
\value{
A tibble with three columns: \code{Detector}, \code{Position}, and \code{Length}.
}
\description{
This function is a helper for the \code{read_melt} function. It extracts the
detectors and their line ranges in the csv file.
}
