% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_CT.R
\name{.set_subset}
\alias{.set_subset}
\title{A helper function to create subsets for calibrator sample and reference assay}
\usage{
.set_subset(data, col, value, name)
}
\arguments{
\item{data}{A tibble with qPCR data.}

\item{col}{A character string. The column to look for matching values.}

\item{name}{A character vector. The name of the created column with boolean
values.}

\item{value.}{A character string. The value to look for matches in column.}
}
\value{
A tibble with an additional column containing boolean values.
}
\description{
This function adds a column called \code{name} with boolean values. It is \code{TRUE}
if the column \code{col} equals the value \code{value} in the dataset \code{data}.
}
