% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{fancy_scientific}
\alias{fancy_scientific}
\title{Change the scientific number style}
\usage{
fancy_scientific(x)
}
\arguments{
\item{x}{A string in scientific notation.}
}
\value{
A string in "fancy" scientific notation.
}
\description{
I take no credit for this function. It was taken from Stackoverflow from
an answer by Jack Aidley:
https://stackoverflow.com/questions/11610377/how-do-i-change-the-formatting-of-numbers-on-an-axis-with-ggplot
The function changes the style of numbers from '0e+0' to '0x10^0'.
}
