% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plates.R
\docType{data}
\name{Plate_384}
\alias{Plate_384}
\title{Layout for 384-well plate}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 384 rows and 4 columns.
}
\usage{
Plate_384
}
\description{
This is a tibble with the coordinates for a 384-well plate.
}
\keyword{datasets}
