% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plates.R
\docType{data}
\name{Plate_96}
\alias{Plate_96}
\title{Layout for 96-well plate}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 4 columns.
}
\usage{
Plate_96
}
\description{
This is a tibble with the coordinates for a 96-well plate.
}
\keyword{datasets}
