% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melting_Analysis.R
\name{read_melt}
\alias{read_melt}
\title{Load melting curves.}
\usage{
read_melt(file_name, plate = "Plate_96")
}
\arguments{
\item{file_name}{The file to load.}

\item{plate}{The plate format. The default is \code{"Plate_96"} but \code{"Plate_384"} is
also acceptable.}
}
\value{
A tibble with the melting data.
}
\description{
This function reads a csv file with melting curves exported from Applied
Biosystems 7500 and 7500 Fast Real-Time PCR System.
}
