% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_CT.R
\name{read_qPCR}
\alias{read_qPCR}
\title{Extract the detectors from a csv files with melting data.}
\usage{
read_qPCR(file, format = "ABI_7500_csv", repair_CT_limits = TRUE)
}
\arguments{
\item{file}{The file to load.}

\item{format}{The file format that is imported. Currently only csv files
from ABI 7500 and ABI 7500 Fast are supported.}

\item{repair_CT_limits}{A bolean value. If \code{TRUE}, it will replace
\code{Undetermined} or \code{NA} values with the next highest out of limit value.}
}
\value{
A tibble with columns named \code{Well}, \code{Sample_Name}, \code{Detector},
\code{Task}, \code{CT}, \code{Std_CT}, \code{Quantity}, \code{Mean_Quantity}, \code{Std_Quantity},
\code{Filtered}, and  \code{Melting_Temp}. The columns can be empty.
}
\description{
This function reads a csv file with qPCR data as exported from ABI 7500.
}
